+++ changelog.md
@@ -1,6 +1,8 @@
 # Changelog
 
-- * 2fd65d3 - 2017-10-30: Refactoring search providers 
+- * 5cea4c4 - 2017-11-02: Fixed issue #102 - Newpct.com changed his web structure 
+- * 7f450e6 - 2017-10-31: Updated grunt file to reflect repo changes 
+- * ebb2a6b - 2017-10-30: Refactoring search providers 
 - * cbb3632 - 2017-10-30: Fixed searches for ThePirateBay using custom urls 
 - * c6bc2e9 - 2017-10-30: Fixed searches for ThePirateBay using custom urls 
 - * 4d52518 - 2017-10-29: Refactoring search providers